requirements.txt
# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.30.0
clip-by-openai==1.1
sentence-transformers>=2.2.0
sentencepiece>=0.1.99
protobuf>=3.20.0
ftfy>=6.1.1
regex>=2022.1.18

# Image Processing
opencv-python>=4.5.0
Pillow>=9.0.0
scikit-image>=0.21.0

# Video Processing
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Audio Processing
whisper-openai>=20230918
librosa>=0.10.0
soundfile>=0.12.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.11.0

# Web and APIs
requests>=2.28.0
aiohttp>=3.8.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=0.19.0
pyyaml>=6.0

# Machine Learning
scikit-learn>=1.0.0
faiss-cpu>=1.7.0  # Use faiss-gpu if CUDA is available
onnxruntime>=1.16.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.17.0

# New dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
python-multipart>=0.0.6
aiofiles>=23.1.0 